% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.GEInfo.R
\name{coef.GEInfo}
\alias{coef.GEInfo}
\title{Extract coefficients from a fitted object}
\usage{
\method{coef}{GEInfo}(object, ...)
}
\arguments{
\item{object}{A fitted "CGEInfo" or "GEInfo" model object for which the estimate of coefficients is extracted.}

\item{...}{Other arguments.}
}
\value{
A coefficient vector of length (q+1) x (p+1), including the estimates for \eqn{\alpha} (intercept), \eqn{a} (coefficients for all E variables), and \eqn{b} (coefficients for all G variables and G-E interactions).
}
\description{
Report the estimate of all coefficients from a fitted "CGEInfo" or "GEInfo" model object.
}
