% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GEInfo.R
\name{plot.GEInfo}
\alias{plot.GEInfo}
\title{Heatmap of the identification results}
\usage{
\method{plot}{GEInfo}(x, Gname = NULL, Ename = NULL, ...)
}
\arguments{
\item{x}{A fitted "GEInfo" model object for which prediction is desired.}

\item{Gname}{Names of all G variables. Default is NULL.}

\item{Ename}{Names of all E variables. Default is NULL.}

\item{...}{Other parameters.}
}
\value{
A Heatmap.
}
\description{
Plot the heatmap for all E variables, identified G variables, and their G-E interactions from a fitted (GEInfo) model.
}
