% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PubMed.R
\name{plot.PubMed}
\alias{plot.PubMed}
\title{Visualize the prior information}
\usage{
\method{plot}{PubMed}(x, G.count = NULL, GE.count = NULL, ...)
}
\arguments{
\item{x}{A 'PubMed' object for which visualization is desired.}

\item{G.count}{A numeric vector of length p, including prior counts for all G variables. Default is NULL.}

\item{GE.count}{A numeric matrix of dimensions p*q, including prior counts for G-E interactions. Default is NULL.}

\item{...}{Other parameters.}
}
\value{
The output includes bar chart for top G variables and G-E interactions by prior counts, and boxplot of prior counts for all G variables.
}
\description{
Visualize the prior counts for G variables and G-E interactions. It reports a bar chart for the top 40 G variables by prior count and
       a boxplot of prior counts for all G variables. For each E variables, it draws a bar chart for the corresponding top 20 G-E interactions by prior count.
}
