% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GEInfo.R
\name{predict.GEInfo}
\alias{predict.GEInfo}
\title{Make Predictions for a fitted model}
\usage{
\method{predict}{GEInfo}(object, Enew, Gnew, family, ...)
}
\arguments{
\item{object}{A fitted "GEInfo" model object for which prediction is desired.}

\item{Enew}{Matrix of dimensions \eqn{n_test x q} for E variables at which predictions are to be made.}

\item{Gnew}{Matrix of dimensions \eqn{n_test x p}for G variables at which predictions are to be made.}

\item{family}{Model type: one of ("gaussian", "binomial", "poisson").}

\item{...}{Other arguments.}
}
\value{
Return a vector of length \eqn{n_test}, representing the fitted response value. For family= “gaussian”, the fitted values are returned;
for family = “binary”, the fitted probabilities are returned;
for family = “poisson”, the fitted means are returned.
}
\description{
Output predicted response values for new observations.
}
